Fix libchdr binding for 32-bit platforms
* Force strip NUL and space when parsing CHD CD metadata tag
* Fix signature of 'fread' callback: use arch-sized integers where necessary
* Force 64-bit alignment for 'FileHeader.LogicalBytes' field

+++ pkg/fs/chd/libchdr_file_callbacks.go
@@ -14,7 +14,7 @@ import (
 	"github.com/xakep666/ps3netsrv-go/internal/osutil/dynamiclibs"
 )
 
-// fileCallbacks are called by libchdr itself to outsource i/o operations
+// fileCallbacks are called by libchdr ituserdata to outsource i/o operations
 type fileCallbacks struct {
 	_ structs.HostLayout
 
@@ -27,6 +27,7 @@ type fileCallbacks struct {
 func newFileCallbacks(log *slog.Logger) *fileCallbacks {
 	// dynamiclibs.Handle used to propagate a Go value through C code without it's being garbage-collected
 	return &fileCallbacks{
+		// signature: uint64_t fsize(void* userdata);
 		fsize: dynamiclibs.NewCallback(func(_ dynamiclibs.CDecl, userdata dynamiclibs.Handle) uint64 {
 			f := userdata.Value().(handler.File)
 			ret, err := f.Seek(0, io.SeekEnd)
@@ -36,19 +37,22 @@ func newFileCallbacks(log *slog.Logger) *fileCallbacks {
 			}
 			return uint64(ret)
 		}),
-		fread: dynamiclibs.NewCallback(func(_ dynamiclibs.CDecl, target *byte, size, count int64, userdata dynamiclibs.Handle) int64 {
+		// signature: size_t fread(void *target, size_t size, size_t count, void* userdata);
+		// Go's analog for size_t is uint.
+		fread: dynamiclibs.NewCallback(func(_ dynamiclibs.CDecl, target *byte, size, count uint, userdata dynamiclibs.Handle) uint {
 			f := userdata.Value().(handler.File)
 			n, err := f.Read(unsafe.Slice(target, count*size))
 			switch {
 			case errors.Is(err, nil):
-				return int64(n)
+				return uint(n)
 			case errors.Is(err, io.EOF):
 				return 0
 			default:
 				log.Error("chd: read failed", logutil.ErrorAttr(err), slog.String("name", f.Name()))
-				return int64(extractSysErrCode(err))
+				return uint(extractSysErrCode(err))
 			}
 		}),
+		// signatrue: int fclose(void *userdata);
 		fclose: dynamiclibs.NewCallback(func(_ dynamiclibs.CDecl, userdata dynamiclibs.Handle) int {
 			defer userdata.Delete()
 
@@ -60,6 +64,7 @@ func newFileCallbacks(log *slog.Logger) *fileCallbacks {
 			}
 			return 0
 		}),
+		// signature: int fseek(void *userdata, int64_t offset, int whence);
 		fseek: dynamiclibs.NewCallback(func(_ dynamiclibs.CDecl, userdata dynamiclibs.Handle, offset int64, whence int) int {
 			f := userdata.Value().(handler.File)
 			_, err := f.Seek(offset, whence)
